........
  r52929 | kwc@umich.edu | 2008-09-09 15:04:32 -0400 (Tue, 09 Sep 2008) | 6 lines
  
  Pieter says that dscribe2s should only see/manage courses to which
  they have been assigned (via an acl table entry).
  Use the correct function to limit the list of courses returned.
  
  Fix the view to allow for an empty course list.
........
  r53004 | kwc@umich.edu | 2008-09-11 11:19:34 -0400 (Thu, 11 Sep 2008) | 3 lines
  
  The notify script contains configuration information,
  so rename it as a template.
........
  r53109 | kwc@umich.edu | 2008-09-15 15:03:37 -0400 (Mon, 15 Sep 2008) | 7 lines
  
  1) Add a new config option and change the FAL_front library code to use it to
     eliminate the hard-coded cosign cookie name.
  
  2) Change instances of "oer" and "oerpath" to something more recognizable
     so they can eventually be changed via a script
  	 or in some other automated fashion.
........
  r53116 | kwc@umich.edu | 2008-09-15 22:10:28 -0400 (Mon, 15 Sep 2008) | 8 lines
  
  An extension to OERDEV-136, this adds the ability to manage users
  for a course.  The administrator can add and remove dscribes
  and instructors for a course.
  
  Do not allow the last user of any category (dscribe1, dscribe2,
  and instructor) to be removed.
........
  r53196 | bdr@umich.edu | 2008-09-17 13:39:37 -0400 (Wed, 17 Sep 2008) | 3 lines
  
  OERDEV-180  bdr - now have progress bars on top of most pages
........
  r53205 | mbleed@umich.edu | 2008-09-17 14:25:08 -0400 (Wed, 17 Sep 2008) | 1 line
  
  change location of co object key
........
  r53209 | dkhutch@umich.edu | 2008-09-17 17:18:36 -0400 (Wed, 17 Sep 2008) | 1 line
  
  OERDEV 181
........
  r53229 | bdr@umich.edu | 2008-09-18 09:01:51 -0400 (Thu, 18 Sep 2008) | 3 lines
  
  OERDEV-182  bdr fixed following Kevin's advice
........
  r53251 | lotia@umich.edu | 2008-09-18 13:45:42 -0400 (Thu, 18 Sep 2008) | 11 lines
  
  Set default CO view to all when no new objects
  
  Jira bug 164.
  
  If there are any content objects without assigned status, then
  the CO view for a particular material shows those "new" objects
  in the initial view when clicking on the link for that material.
  
  However, if all objects for a material have assigned status which
  means there are no "new" objects, then all objects are visible in
  the initial view (All) when clicking on the link for that material.
........
  r53257 | mbleed@umich.edu | 2008-09-18 16:07:45 -0400 (Thu, 18 Sep 2008) | 1 line
  
  OERDEV-183 fix autoscroll
........
  r53258 | mbleed@umich.edu | 2008-09-18 16:42:01 -0400 (Thu, 18 Sep 2008) | 1 line
  
  OERDEV-168 set default view on instructor ask forms
........
  r53259 | lotia@umich.edu | 2008-09-18 16:47:46 -0400 (Thu, 18 Sep 2008) | 7 lines
  
  CO display was being forced to "New"
  
  The fix for OERDEV-164 was incorrect. The CO display was being forced to
  display the "New" view regardless of what was selected. This fix changes
  that so that if no view is specified and no new COs are present, it
  displays the "All" view and when a view is specified it displays that
  view.
........
  r53261 | kwc@umich.edu | 2008-09-18 17:12:22 -0400 (Thu, 18 Sep 2008) | 12 lines
  
  OERDEV-178
  
  Allow a custom prompt in the confirmation dialogs when the admin
  is deleting users, curriculums, subjects, schools, or courses.
  
  Changes the 'confirm' Javascript class to look for a 'customprompt'
  attribute.  If none was specified, fall back to the default prompt,
  "Are you sure?".
  
  Change admin uses of the confirm class to supply custom prompts.
........
  r53262 | lotia@umich.edu | 2008-09-18 17:38:00 -0400 (Thu, 18 Sep 2008) | 6 lines
  
  Retrieve and display COs ordered by location
  
  OERDEV-165. Content objects were previously displayed by date of addition
  or modification, they are now retrieved by their page location and displayed
  accordingly. This ordering may be a little buggy since it relies on casting
  the varchar 'location' field to unsigned, but I haven't broken it yet.
........
  r53264 | kwc@umich.edu | 2008-09-18 18:27:28 -0400 (Thu, 18 Sep 2008) | 6 lines
  
  OERDEV-172
  
  If there are no Replacment COs, flash a message and reload
  the page rather than winding up with a blank page.
........
  r53272 | kwc@umich.edu | 2008-09-19 11:33:07 -0400 (Fri, 19 Sep 2008) | 8 lines
  
  OERDEV-178
  
  Add more custom prompts for deletion of Content Objects,
  Replacement COs, and Materials.
  
  Also for admin user management. (Unassigning users from
  coursea,s or dscribe1s from dscribe2s.)
........
  r53273 | lotia@umich.edu | 2008-09-19 15:22:29 -0400 (Fri, 19 Sep 2008) | 7 lines
  
  Set Embedded COs "Yes" for bulk materials uploads
  
  OERDEV-186
  When materials bulk uploaded, the embedded CO value was previously set
  to "No" on Pieter's request, it is now set to "Yes". The fix could have
  been made through and update to the DB schema or a hidden form field.
  This fix uses the latter approach.
........
  r53277 | kwc@umich.edu | 2008-09-19 16:06:59 -0400 (Fri, 19 Sep 2008) | 4 lines
  
  OERDEV-187
  
  Update the 'modified_on' for a material any time something in the material is updated.
........
  r53278 | kwc@umich.edu | 2008-09-19 16:14:19 -0400 (Fri, 19 Sep 2008) | 3 lines
  
  Per Piet's request, revert revision 53257 which was
  made in response to OERDEV-183.
........
  r53279 | kwc@umich.edu | 2008-09-19 16:17:18 -0400 (Fri, 19 Sep 2008) | 2 lines
  
  Arrgghh!  Remove debugging accidently checked in with revision 53277.
........

+++ tool/system/application/models/course.php
@@ -79,14 +79,111 @@ class Course extends Model
 	}
 	
 	/**
+		* Get users (acl entries) associated with a course
+		*
+		* @access  public
+		*	@param   int course_id
+		*	@param   string user role (optional)
+		*	@return  array of ocw_acl entries
+		*/
+	public function get_course_users($cid, $role = NULL) {
+		$where = ($role == NULL) ? array('course_id' => $cid)
+														 : array('course_id' => $cid, 'role' => $role);
+		$this->db->select('*')->from('acl')->where($where);
+		$q = $this->db->get();
+		if ($q->num_rows() > 0) {
+			foreach($q->result_array() as $row) {
+				$cacls[] = $row;
+    	}
+		}
+		return ($q->num_rows() > 0) ? $cacls : null;
+	}
+
+	/**
+    * Complete (dscribe2_dscribe1) relationships when adding an acl entry
+		*
+		* @access  private
+		* @param   array details (containing course_id, user_id, and role)
+		* @return  boolean
+		*/
+	private function complete_relationships($details)
+	{
+		if ($details['role'] != 'dscribe1' && $details['role'] != 'dscribe2')
+			return true;
+		if ($details['role'] == 'dscribe1') {
+			// create a relationship with existing dscribe2s
+			$d2s = $this->get_course_users($details['course_id'], 'dscribe2');
+			if (count($d2s) != 0) {
+				foreach($d2s as $d2) {
+					$this->ocw_user->set_relationship($details['user_id'], $d2['user_id']);
+				}
+			}
+		} else if ($details['role'] == 'dscribe2') {
+			// create a relationship with existing dscribe1s
+			$d1s = $this->get_course_users($details['course_id'], 'dscribe1');
+			if (count($d1s) != 0) {
+				foreach($d1s as $d1) {
+					$this->ocw_user->set_relationship($d1['user_id'], $details['user_id']);
+				}
+			}
+		}
+		return true;
+	}
+
+	/**
+		* This function checks all the existing acl rows and verifies that
+		* the role claimed in that row matches the roll of the user pointed
+		* to by the user_id field
+		* Inconsistencies are logged to the apache error log
+		*
+	  * @access private
+	  * @return true
+	  */
+	private function check_for_inconsistent_acls()
+	{
+		$this->db->select("*")->from('acl');
+		$q = $this->db->get();
+		if ($q->num_rows() > 0) {
+			foreach($q->result_array() as $acl) {
+				$this->db->select('id, role')->from('users')->where(array('id' => $acl['user_id']))->limit(1);
+				$qu = $this->db->get();
+				if ($qu->num_rows() != 1) {
+					$errmsg = "check_for_inconsistent_acls: Invalid acl entry "
+								. "(user_id " . $acl['user_id']
+								. ", course_id " . $acl['course_id']
+								. ", role " . $acl['role']
+								. "), no matching user entry found";
+					$this->ocw_utils->log_to_apache('warn', $errmsg);
+				}
+				foreach($qu->result_array() as $u) {
+					if ($u['role'] != $acl['role']) {
+						$errmsg = "check_for_inconsistent_acls: Invalid acl entry "
+									. "(user_id " . $acl['user_id']
+									. ", course_id " . $acl['course_id']
+									. ", role " . $acl['role']
+									. "), role doesn't match user entry role (" . $u['role'] . ")";
+						$this->ocw_utils->log_to_apache('warn', $errmsg);
+						// This goes to the CodeIgniter logs: log_message('error', $errmsg);
+						// This goes to the php_error.log: error_log($errmsg, 0);
+					}
+				}
+			}
+		}
+		return true;
+	}
+
+	/**
      * add user with the role to the course
 	 *
 	 * @access  public
+	 * @param   array details (containing course_id, user_id, and role)
 	 * @return  boolean
 	 */
   public function add_user($details)
   {
+		$this->complete_relationships($details);
 		$this->db->insert('acl', $details);
+		$this->check_for_inconsistent_acls();
 		return true;
 	}
 
@@ -101,6 +198,21 @@ class Course extends Model
     */
   public function remove_user($cid, $uid, $role)
   {
+		// Don't allow deletion of the last user in any particular
+		// role (dscribe1, dscribe2, or instructor) from a course
+		$d1s = $this->get_course_users($cid, 'dscribe1');
+		$d2s = $this->get_course_users($cid, 'dscribe2');
+		$instrs = $this->get_course_users($cid, 'instructor');
+
+		if ($role == 'instructor' && count($instrs) <= 1)
+			return "Cannot remove the only instructor from a course";
+		if ($role == 'dscribe2' && count($d2s) <= 1)
+			return "Cannot remove the only dscribe2 from a course";
+		if ($role == 'dscribe1' && count($d1s) <= 1)
+			return "Cannot remove the only dscribe1 from a course";
+		if ((count($d2s) + count($d1s) + count($instrs)) <= 1)
+			return "Cannot remove the only user for a course";
+		
 		$d = array('user_id'=>$uid, 'course_id'=>$cid, 'role'=>$role);
 		$this->db->delete('acl',$d);
 		return true;
